Fix:デフォルト画像配置場所

+++ app/models/user.rb
@@ -36,17 +36,18 @@ class User < ApplicationRecord
     if profile_image.attached?
       profile_image.variant(resize_to_fill: [width, height]).processed
     else
-      file_path = Rails.root.join('app/assets/images/no_image_square.jpg')
-      file = File.open(file_path)
+      file_path = Rails.root.join('public/images/no_image_square.jpg')
+      return unless File.exist?(file_path) # 念のためファイル存在確認
   
-      dummy_blob = ActiveStorage::Blob.create_and_upload!(
+      file = File.open(file_path)
+      blob = ActiveStorage::Blob.create_and_upload!(
         io: file,
         filename: 'no_image_square.jpg',
         content_type: 'image/jpeg'
       )
       file.close
   
-      dummy_blob.variant(resize_to_fill: [width, height]).processed
+      blob.variant(resize_to_fill: [width, height]).processed
     end
   end
